gccrs: backend: Expose Bvariable class through rust-gcc header
gcc/rust/ChangeLog:

	* rust-gcc.cc (class Bvariable): Move class to `rust-gcc.h` header.
	* rust-gcc.h: New file.
